+++ Builds/Build Number 0/Build 0.log
@@ -1,0 +1,22 @@
+[2024-07-03 00:02:38,153 - INFO - [Executar] Iniciando execução [0] datetime [2024-07-03 00:02:38]]
+[2024-07-03 00:02:44,084 - WARNING - [ConsorcioBB] Versão Selenium Em Execução (V4.21.0).]
+[2024-07-03 00:02:44,084 - INFO - [ConsorcioBB] Iniciando Chrome...]
+[2024-07-03 00:02:46,995 - INFO - [ConsorcioBB] Acessando URL (https://www.parceirosbbconsorcios.com.br).]
+[2024-07-03 00:02:47,662 - INFO - [ConsorcioBB] Efetuando login PP BB Consórcio.]
+[2024-07-03 00:02:48,028 - INFO - [ConsorcioBB] Logado com sucesso no PP BB Consórcio.]
+[2024-07-03 00:02:48,029 - INFO - [ConsorcioBB] Fechando Poup-Up Inicial.]
+[2024-07-03 00:02:48,534 - INFO - [ConsorcioBB] Poup-Up fechado.]
+[2024-07-03 00:03:06,902 - INFO - [ConsorcioBB] Data frame Grupos Ativos [TC] extraído com sucesso. ]
+[2024-07-03 00:03:30,346 - INFO - [ConsorcioBB] Data frame Grupos Ativos [AI] extraído com sucesso. ]
+[2024-07-03 00:04:15,146 - INFO - [ConsorcioBB] Data frame Grupos Ativos [AU] extraído com sucesso. ]
+[2024-07-03 00:04:31,250 - INFO - [ConsorcioBB] Data frame Grupos Ativos [MO] extraído com sucesso. ]
+[2024-07-03 00:04:54,385 - INFO - [ConsorcioBB] Data frame Grupos Ativos [EE] extraído com sucesso. ]
+[2024-07-03 00:05:06,270 - INFO - [ConsorcioBB] Data frame Grupos Ativos [IM240] extraído com sucesso. ]
+[2024-07-03 00:05:09,054 - INFO - [ConsorcioBB] Data frame Grupos Ativos [IMP] extraído com sucesso. ]
+[2024-07-03 00:05:12,250 - INFO - [Executar] Arquivo 'df_todosGruposAtivos.xlsx' criado com sucesso.]
+[2024-07-03 00:06:21,401 - INFO - [ConsorcioBB] Dados das Assembleias Extraídos com Sucesso.]
+[2024-07-03 00:06:21,749 - INFO - [Executar] Arquivo 'df_dados_assembleia.xlsx' criado com sucesso.]
+[2024-07-03 01:07:58,090 - INFO - [ConsorcioBB] Média contemplações extraídas com sucesso.]
+[2024-07-03 01:07:58,266 - INFO - [Executar] Arquivo 'df_medias_assembleias.xlsx' criado com sucesso.]
+[2024-07-03 01:07:58,745 - INFO - [Executar] Arquivo 'TodosGruposAtivos.xlsx' criado com sucesso.]
+[2024-07-03 01:07:58,745 - INFO - [ConsorcioBB] Encerrando Chrome.]
